requirements.txt
spacecases-common
discord.py
python-dotenv
asyncpg
requests
mypy
types-requests
marisa-trie
